In the process of figuring out how to use fakedatabase
THIS COMMIT IS BROKEN
I have a general idea of how the database will work, but i need to
figure out how provider will fit into it. Additionally, the servlets and
JSPs are going to need to carry the FakeDatabase up and down the HTTP
requests so that it persists, if you make a new database each time you
save, you're not actually saving anything.